backend/requirements.txt
flask==3.0.3
flask-sqlalchemy==3.1.1
flask-migrate==4.0.5
flask-jwt-extended==4.6.0
flask-cors==4.0.1
mysql-connector-python==9.0.0
python-dotenv==1.0.1
werkzeug==3.0.3













